Move TableCell to own file
