+++ frontend/src/metabase/visualizations/components/TableSimple/TableSimple.jsx
@@ -11,23 +11,16 @@ import { getIn } from "icepick";
 import _ from "underscore";
 import { t } from "ttag";
 
-import ExternalLink from "metabase/core/components/ExternalLink";
 import ExplicitSize from "metabase/components/ExplicitSize";
 import Ellipsified from "metabase/core/components/Ellipsified";
 import Icon from "metabase/components/Icon";
 
-import { formatValue } from "metabase/lib/formatting";
 import { isPositiveInteger } from "metabase/lib/number";
-import { isID, isFK } from "metabase/lib/schema_metadata";
+import { isID } from "metabase/lib/schema_metadata";
 import { HARD_ROW_LIMIT } from "metabase/lib/query";
-import {
-  getTableCellClickedObject,
-  getTableClickedObjectRowData,
-  isColumnRightAligned,
-} from "metabase/visualizations/lib/table";
-import { getColumnExtent } from "metabase/visualizations/lib/utils";
+import { isColumnRightAligned } from "metabase/visualizations/lib/table";
 
-import MiniBar from "../MiniBar";
+import TableCell from "./TableCell";
 import styles from "./Table.css";
 
 function getBoundingClientRectSafe(ref) {
@@ -53,156 +46,6 @@ const CELL_HEADER_ICON_STYLE = {
   right: "100%",
   marginRight: 3,
 };
-
-function getCellData({
-  value,
-  clicked,
-  extraData,
-  cols,
-  rows,
-  columnIndex,
-  columnSettings,
-}) {
-  if (value == null) {
-    return "-";
-  }
-  if (columnSettings["show_mini_bar"]) {
-    return (
-      <MiniBar
-        value={value}
-        options={columnSettings}
-        extent={getColumnExtent(cols, rows, columnIndex)}
-      />
-    );
-  }
-  return formatValue(value, {
-    ...columnSettings,
-    clicked: { ...clicked, extraData },
-    type: "cell",
-    jsx: true,
-    rich: true,
-  });
-}
-
-function TableCell({
-  value,
-  data,
-  series,
-  settings,
-  rowIndex,
-  columnIndex,
-  isPivoted,
-  getCellBackgroundColor,
-  getExtraDataForClick,
-  checkIsVisualizationClickable,
-  onVisualizationClick,
-}) {
-  const { rows, cols } = data;
-  const column = cols[columnIndex];
-  const columnSettings = settings.column(column);
-
-  const clickedRowData = useMemo(
-    () =>
-      getTableClickedObjectRowData(
-        series,
-        rowIndex,
-        columnIndex,
-        isPivoted,
-        data,
-      ),
-    [data, series, rowIndex, columnIndex, isPivoted],
-  );
-
-  const clicked = useMemo(
-    () =>
-      getTableCellClickedObject(
-        data,
-        settings,
-        rowIndex,
-        columnIndex,
-        isPivoted,
-        clickedRowData,
-      ),
-    [data, settings, rowIndex, columnIndex, isPivoted, clickedRowData],
-  );
-
-  const extraData = useMemo(() => getExtraDataForClick?.(clicked) ?? {}, [
-    clicked,
-    getExtraDataForClick,
-  ]);
-
-  const cellData = useMemo(
-    () =>
-      getCellData({
-        value,
-        clicked,
-        extraData,
-        cols,
-        rows,
-        columnIndex,
-        columnSettings,
-      }),
-    [value, clicked, extraData, cols, rows, columnIndex, columnSettings],
-  );
-
-  const isLink = cellData && cellData.type === ExternalLink;
-  const isClickable = !isLink && checkIsVisualizationClickable(clicked);
-
-  const onClick = useMemo(() => {
-    if (!isClickable) {
-      return;
-    }
-    return e => {
-      onVisualizationClick({
-        ...clicked,
-        element: e.currentTarget,
-        extraData,
-      });
-    };
-  }, [isClickable, clicked, extraData, onVisualizationClick]);
-
-  const style = useMemo(() => {
-    const result = { whiteSpace: "nowrap" };
-    if (getCellBackgroundColor) {
-      result.backgroundColor = getCellBackgroundColor(
-        value,
-        rowIndex,
-        column.name,
-      );
-    }
-    return result;
-  }, [value, rowIndex, column, getCellBackgroundColor]);
-
-  const classNames = useMemo(
-    () =>
-      cx(
-        "px1 border-bottom text-dark fullscreen-normal-text fullscreen-night-text text-bold",
-        {
-          "text-right": isColumnRightAligned(column),
-          "Table-ID": value != null && isID(column),
-          "Table-FK": value != null && isFK(column),
-          link: isClickable && isID(column),
-        },
-      ),
-    [value, column, isClickable],
-  );
-
-  const classNames2 = useMemo(
-    () =>
-      cx("cellData inline-block", {
-        "cursor-pointer text-brand-hover": isClickable,
-      }),
-    [isClickable],
-  );
-
-  return (
-    <td className={classNames} style={style}>
-      <span className={classNames2} onClick={onClick}>
-        {cellData}
-      </span>
-    </td>
-  );
-}
 
 function TableSimple({
   card,
